FASTAPI/api_detections/requirements.txt
# Base YOLO ------------------------------------------------------------------------
gitpython>=3.1.30
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
psutil  # system resources
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
thop>=0.1.1  # FLOPs computation
tqdm>=4.64.0
ultralytics>=8.0.111
# protobuf<=3.20.1  # https://github.com/ultralytics/yolov5/issues/8012
# Plotting --------------------------------------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# DJANGO
fastapi
uvicorn[standard]

# ???
albumentations
pyshine
python-gettext

# DeepSort
deep-sort-realtime

# aiocache
aiocache
redis
jsonpickle